requirements/constraints.txt
black==24.4.2
eyed3==0.9.7
faker==33.1.0
FATtools==1.0.39
isort==5.13.2
multipledispatch==1.0.0
mypy==1.10.0
pre-commit==3.8.0
pydantic==2.10.4
pyfakefs==5.7.3
pytest==8.3.4
pytest-cov==5.0.0
pytest-env==1.1.3
pytest-mock==3.14.0
pytest-mockito==0.0.4
pytest-timeout==2.3.1
pyyaml==6.0.2
strip_ansi==0.1.1
types-PyYAML==6.0.12.20240917
typer==0.15.1
wemake-python-styleguide==0.19.2
wheel==0.45.1
